requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.60.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.0.4
et_xmlfile==2.0.0
fastapi==0.104.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
ida-pro-mcp @ git+https://github.com/mrexodia/ida-pro-mcp@f7ea8d66a31fe6b402a5c234c2912c351bacdf15
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.26
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.1
litellm==1.73.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.13.1
mcp-server-ida==0.3.4
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.1
ollama==0.5.1
openai==1.90.0
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.3
paramiko==3.5.1
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
pyparsing==3.2.4
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.6
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.41
sse-starlette==3.0.2
starlette==0.27.0
tavily-python==0.7.10
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==12.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
